GenAI SmartOps Integrated Platform/requirements.txt
langchain
langchain_community
langchain_core
langgraph
chroma
gradio
openai
faiss-cpu 
transformers
torch
pandas
numpy
scipy
matplotlib
seaborn
scikit-learn
requests
psutil
ansible
grafana-client
python-dotenv
json5
looging
re
random
requests
webbrowser
gradio
flask
requests
python-dotenv
tqdm
watchdog
time
json
subprocess
logging
jsonify